refactor to use re-export of components with styles
