+++ app/src/components/tag/index.tsx
@@ -1,0 +1,3 @@
+export * from "./Tag";
+export * from "./TagGroup";
+export * from "./TagList";
